Fix ResolverEventLogger thread leak
by letting the logging thread die when no events to process for a while
and restarting it when required

One event might hang in the queue if it arrives after q.poll()
has returned null and before the consumer thread death.
In this case, the next event will flush the queue to disk.